E2E: visualize image + run through model
This extends the notebook with 2 features:

1) Render test data into an image

This seemed like a quick way to get a .png file from the 0-1 pixel representation of numbers.

2) Predict image using model

Given a png file, convert it to 0-1 normalized pixels, then run it through the model.

TODO: what if it's not grayscale?

----

This wraps up an e2e proof of concept of the overall flow: training the model from data, then using the model to predict received pixels.