Add cloud function to classify base64 image
This cloud function receives a base64 image. The content is a JSON
object containing one field, "image_b64", equal to the base64 encoding
of an image file. I think it has to be 28x28 grayscale (single channel).

The deploy.sh script wraps the Google CLI to deploy the function with
appropriate environment variables.